+++ src/TMModels/League.cs
@@ -20,9 +20,9 @@ public record League(
 }
 
 public record Scoring(
-    [property: JsonPropertyName("pointsPerStronghold")] ushort PointsPerStronghold,
-    [property: JsonPropertyName("pointsPerCastle")] ushort PointsPerCastle,
-    [property: JsonPropertyName("pointsPerWin")] ushort PointsPerWin,
+    [property: JsonPropertyName("pointsPerStronghold")] decimal PointsPerStronghold,
+    [property: JsonPropertyName("pointsPerCastle")] decimal PointsPerCastle,
+    [property: JsonPropertyName("pointsPerWin")] decimal PointsPerWin,
     [property: JsonPropertyName("requiredBattlesBefore10thTurn")] ushort RequiredBattlesBefore10thTurn,
     [property: JsonPropertyName("tiebreakers")] Tiebreaker[] Tiebreakers
 );
